pip-requirements.txt
jsonpickle
geojson
shapely
GeoAlchemy>=0.6
requests==1.1.0
-e git+https://github.com/ckan/ckanext-spatial.git#egg=ckanext_spatial
zope.interface==4.0.1
zope.schema==4.4
zope.component==4.1
zope.dottedname
z3c.schema
owslib==0.8.8
